breaking: sort by using index name only
