+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/sortby/SortByConnection.kt
@@ -6,8 +6,8 @@ import com.algolia.instantsearch.compose.sortby.internal.SortByConnectionPaginat
 import com.algolia.instantsearch.core.connection.Connection
 import com.algolia.instantsearch.core.searcher.Searcher
 import com.algolia.instantsearch.helper.searcher.IndexNameHolder
-import com.algolia.instantsearch.helper.sortby.searcher.SortByConnector
-import com.algolia.instantsearch.helper.sortby.searcher.SortByViewModel
+import com.algolia.instantsearch.helper.sortby.SortByConnector
+import com.algolia.instantsearch.helper.sortby.SortByViewModel
 
 
 /**
