+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/sortby/internal/SortByConnectionPaginator.kt
@@ -4,7 +4,7 @@ import com.algolia.instantsearch.ExperimentalInstantSearch
 import com.algolia.instantsearch.compose.list.Paginator
 import com.algolia.instantsearch.core.Callback
 import com.algolia.instantsearch.core.connection.ConnectionImpl
-import com.algolia.instantsearch.helper.sortby.searcher.SortByViewModel
+import com.algolia.instantsearch.helper.sortby.SortByViewModel
 
 /**
  * Connection between [Paginator] and [SortByViewModel].
