+++ instantsearch/src/commonMain/kotlin/com/algolia/instantsearch/helper/searcher/FilterGroupsHolder.kt
@@ -1,12 +1,10 @@
 package com.algolia.instantsearch.helper.searcher
 
-import com.algolia.instantsearch.ExperimentalInstantSearch
 import com.algolia.search.model.filter.FilterGroup
 
 /**
  * Component holding filter groups set.
  */
-@ExperimentalInstantSearch
 public interface FilterGroupsHolder {
 
     /**
