+++ instantsearch/src/commonMain/kotlin/com/algolia/instantsearch/helper/sortby/SortByView.kt
@@ -2,4 +2,7 @@ package com.algolia.instantsearch.helper.sortby
 
 import com.algolia.instantsearch.core.selectable.map.SelectableMapView
 
+/**
+ * Sort By view.
+ */
 public typealias SortByView = SelectableMapView<Int, String>
