+++ instantsearch/src/commonMain/kotlin/com/algolia/instantsearch/helper/sortby/internal/SortByConnectionSearcher.kt
@@ -2,14 +2,17 @@ package com.algolia.instantsearch.helper.sortby.internal
 
 import com.algolia.instantsearch.core.Callback
 import com.algolia.instantsearch.core.connection.ConnectionImpl
-import com.algolia.instantsearch.helper.searcher.SearcherIndex
+import com.algolia.instantsearch.core.searcher.Searcher
+import com.algolia.instantsearch.helper.searcher.IndexNameHolder
 import com.algolia.instantsearch.helper.sortby.SortByViewModel
 
-@Deprecated("use SortByConnectionSearcher with generic searcher type instead")
-internal data class SortByConnectionSearcher(
+/**
+ * Sort by widget connection to a searcher.
+ */
+internal data class SortByConnectionSearcher<S>(
     private val viewModel: SortByViewModel,
-    private val searcher: SearcherIndex<*>,
-) : ConnectionImpl() {
+    private val searcher: S,
+) : ConnectionImpl() where S : Searcher<*>, S : IndexNameHolder {
 
     init {
         viewModel.updateSelection()
@@ -31,6 +34,6 @@ internal data class SortByConnectionSearcher(
     }
 
     private fun SortByViewModel.updateSelection(selection: Int? = selected.value) {
-        map.value[selection]?.let { searcher.index = it }
+        map.value[selection]?.let { searcher.indexName = it }
     }
 }
